requirements.txt
python-dotenv
openai
varname